refactoring done with formatting fixes
