+++ backend/app/__init__.py
@@ -26,7 +26,13 @@ def create_app(config_name):
 
     # TODO: Get the router from the main file and other files later on
     from app.main import main
+    from app.main.routers import food, food_collect, vendors, session, trays
 
     app.include_router(main.router)
+    app.include_router(food.router)
+    app.include_router(food_collect.router)
+    app.include_router(vendors.router)
+    app.include_router(session.router)
+    app.include_router(trays.router)
 
     return app
